feat(proposal): fix circular dependencies
