--- packages/sdk-react/src/components/userRoles/UserRoleDetails/UserRoleDetailsDialog/UserRoleRow/index.ts
@@ -1,1 +1,0 @@
-export { UserRoleRow } from './UserRoleRow';
