+++ packages/sdk-react/src/components/userRoles/UserRoleDetails/UserRoleDetailsDialog/const.ts
@@ -1,0 +1,8 @@
+/** View of the user role details */
+export enum UserRoleDetailsView {
+  /** Read mode - the user is only viewing the role details */
+  Read = 'read',
+
+  /** Mutate mode - the user is adding or editing the role */
+  Mutate = 'mutate',
+}
