+++ packages/sdk-react/src/components/userRoles/UserRoleDetails/UserRoleDetailsDialog/index.ts
@@ -1,1 +1,3 @@
-export * from './UserRoleDetailsDialog';
+export { UserRoleDetailsDialog } from './UserRoleDetailsDialog';
+export { UserRoleDetailsView } from '@/components/userRoles/UserRoleDetails/UserRoleDetailsDialog/const';
+export { StyledUserRoleTableCell } from '@/components/userRoles/UserRoleDetails/UserRoleDetailsDialog/StyledUserRoleTableCell';
