+++ packages/sdk-react/src/components/userRoles/UserRoles.tsx
@@ -2,7 +2,8 @@
 
 import React, { useState } from 'react';
 
-import { Dialog, PageHeader, UserRoleDetails } from '@/components';
+import { Dialog } from '@/components/Dialog';
+import { PageHeader } from '@/components/PageHeader';
 import { MoniteScopedProviders } from '@/core/context/MoniteScopedProviders';
 import { useEntityUserByAuthToken } from '@/core/queries';
 import { useIsActionAllowed } from '@/core/queries/usePermissions';
@@ -12,7 +13,8 @@ import { useLingui } from '@lingui/react';
 import { ActionEnum } from '@monite/sdk-api';
 import { Button } from '@mui/material';
 
-import { UserRolesTable } from './UserRolesTable';
+import { UserRoleDetails } from './UserRoleDetails/UserRoleDetails';
+import { UserRolesTable } from './UserRolesTable/UserRolesTable';
 
 export const UserRoles = () => (
   <MoniteScopedProviders>
